refactor(Collections.Generic): collapse DequeBase<T> into Deque<T>
DequeBase<T> existed as a hypothetical extension point for additional
double-ended ring-backed collections, but no second implementation ever
materialised, Deque<T> is sealed, and the type appeared in zero public
APIs as a parameter or return type. Sibling collections (CircularBuffer<T>,
EvictingDictionary<T>) inherit directly from their underlying primitives
without an intermediate family base, so DequeBase<T> was the lone instance
of this two-tier pattern.

The shared abstraction is RingBackedCollection<T>, not DequeBase<T>. Move
RemoveFirst/RemoveLast/PeekFirst/PeekLast and their Try* variants directly
onto Deque<T>, drop the abstract Add* declarations and the `override`
keyword on Deque<T>'s implementations, and reparent Deque<T> to
RingBackedCollection<T>. Update two cross-reference doc comments in
RingBackedCollection.cs and delete DequeBase.cs.

No behaviour change. The only public-API delta is the removal of the
unused public abstract DequeBase<T> type.
